requirements.txt
aldjemy==0.10.1
Django==2.0
django-crispy-forms==1.9.0
djangorestframework==3.11.0
psycopg2==2.8.4
pytz==2019.3
SQLAlchemy==1.3.13
